sort: isolate WASI sync code into sibling modules
